RAG/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0

openai==0.27.8
sentence-transformers==2.2.2
faiss-cpu==1.7.4

torch>=1.9.0
transformers==4.31.0

pydantic==1.10.8
